Rename environment to mirror discrete nature
